Begin implementing akka FSM (seems nice)
Note: does not compile yet. Also, should probably reimplement
MessageFlowTransformer so that WebSocketActor receives and sends
Message:s instead of JsValue:s - makes for cleaner code. Of course
the downside is that then web socket clients will not receive an
explanation if they send json that cannot be parsed as Message...
Something to think about...