Add sql script for mysql db, begin structuring different dbs
I will need to be able to specify different databases for dev
(H2) and production (MySql). However, they accept slightly different
SQL syntax. H2 needs an alias for uuid and has different syntax for
default timestamps. MySQL only supports constants as default values
therefore that configuration requires a "trigger", etc.

So what needs to be done is to have two differently named databases
so that I can make two different sql scripts for evolutions. I must
be able to tell the SlickUserRepository to use one db or the other
depending on if in dev mode etc. This should be possible to achieve
with DI but is still a bit tricky.