Improve environment configuration: test, dev, production
This (sort of) resolves issues #1, #2

Unfortunately the dev database that is used in application.conf, has
evolutions that are applied even if one used test (or prod) conf files.
That means that I cannot (for some reason) use file-based H2 in dev mode.
Also, there is really no dev mode or prod mode. Rather, the tests auto
apply test mode so that conf file (application.test.conf) is automatically
loaded. But that does include "application" which contains the dev specific
database configuration. The alternative is of course to not apply that in
the main application.conf but that would mean having to always pass in
-Dconfig.file="conf/application.dev.conf" flag to sbt because apparently there
is no other way of telling it in build.sbt for instance based on the env automatically.
At least no way that is elegant.