Refactor some logic of WebSocketActor to ListActor
All of the previous functionality still works except ALL().
The reason is that ListActor does not wrap the items in a Response
which would not be possible at this moment. So Response should be
made a sealed trait with two members: either a normal StatusReponse
(what Response was previously) or Items which carries the items.