Fix tests, add smarter testing of WebSocketActor w. ADD_ITEM test
Now it is straight-forward to expect messages from mockWsActor which
now is implemented as TestProbe allowing me to do expectMsg(500 millis,
some_message) or if I need to do more advanced comparisons just
receiveOne(500 millis) - really nice. The problem earlier was that
I could not ask mockWsActor for this since the fsm's ascynchronous
workings made it all rather difficult (the message I expected would
not have arrived when I did mockWsActor ? GetMessage.

Next is to add tests for more of the logic in the fsm, which is rather
sweet! Now just need to test multiple mockWsActor...