Play more with debian:packageBin
Some issues w. db connection when using application.prod.conf, not sure
why if I start it with "listan-server" in the VM it seems to work fine.
But with "initctl listan-server" it does not seem to read the
$DB_PASSWORD env variable (and then probably not $CRYPTO_SECRET either,
but of cours it doesn't come that far since it dies when trying to
connect to the DB, saying using password NO). However, I got it to
connect by purging the DB but that shouldn't be necessary since the
play_evolutions table should keep everything "in sync". An alternative
of course is to not perform the evolutions by default...

TODO:
* investigate further play.pid and /dev/null etc.
* write a note about deployment further in README (upstart scripts is
 automatically handled, but not systemd, and one needs to add env
 variables definitions in /etc/default/listan-server but of course this
 is expected behaviour, but unfortunatly does not work properly...)