Fix extremely annoying jwt bug (string with quotes)
The /api/login action returns a Json body with the token as the string
only which is perfectly valid json.
But previously the test used the header "Authorization Bearer <token>"
to get the token, which worked fine. But for some reason that header is
no longer allowed so I removed it and converted the test to use the
token from the body as "contentAsString(res)" but that maintained the
quotes! So I should have used contentAsJson(res).as[String] which is
what I use now! This took a long time to fix!!! Probably should not
return a Json response if the only content is a String but rather a
plaintext response with the token inside or alternatively a more usual
looking json: {"token": "ey..ms.dfm.m"}...

/End of rant