Implement GetStateResponse with itemListsByUser
Users sending a `GetState` message will only receive lists belonging to
them. However, users that have somehow gotten a hold of an item's uuid
can still read, update and delete it, even if it belongs to another
user. This is because the listActor/webSocketActor does not
authorization based on users apart from GetState where it fetches
itemlists by user uuid as this is enforced in the service layer.
This should conclude the new multiple lists features.