Use Map[ActorRef, UserName] for `clients` in listActor
This allows better handoff of user name infor from websocketator to
listActor as it is done when subscribing to the listActor instead of
manually monkey-patching messages from ws before sending on to listActor
for each action that listActor requires user knowledge of.

Next step would perhaps be to lookup the actual user from db (and thus
the uuid) when transitioning or before, thus ensuring that uuid is valid
and in the database and not just rely on the token. Then the cost of
authorizing actions would decrease as one would already know the user
uuid when performing an action. Would still need to actually figure out
if a certain user is authorized to e.g. edit a specific item of though.