wip: Refactor subscription of WebSocketActors to ListActor
Implement GetStateResponse in ListActor.
Now there will be a third state `PendingSubscription` after
WebSocketActor has authenticated via jwt, ListActor receives the
subscribe message and looks up the userName in the database to retrieve
the user_uuid which is later used for e.g. AddList and GetStateResponse,
during this user retrieval which is async, the ws-actor is not yet
subscribed, and cannot yet do anything. When the async user retrieval
returns, listActor will send a response SubscriptionSucceeeded or
SubscriptionFailed, after which the ws-actor will eiter be added to the
client list and become fully authenticated, or be denied access. Which
can happen in the case that the jwt is valid but the user is not in the
database.

Generally the idea is to trust the jwt token (it can be revoked in any
case and/or expired) but the jwt does not currently store the user uuid
which is needed in some cases. So the solution to that problem could be
to either store that in the jwt which could be maybe not bad practice
but not ideal or do it this way - i.e. look up the user uuid when the
ws-actor subscribes, which is not ideal either and does introduce some
complexity to the app... I should probably think this through further.
/END OF RANT