Rename docker-compose services, remove mysql user uuid trigger
The trigger overwrote any uuid sent in by the client, causing foreign
key constraints to kick in when the client though it referred to the
uuid it just sent in, but in fact a completely different uuid was in
used by the database.
This also means that I should probably just switch to using db in
dev and production and also switch to postgres.