wip: Fix default values in db not being respected by Slick
The problem has been found! When providing a uuid, I did a forceInsert
which overrides the O.AutoInc columns thus inserting null and not
respecting default values for the databse. The difference between
postgres and mysql is that mysql seems to prefer using the default value
instead of inserting nulls while postgres actually sets it to null,
overriding the specified default. So the only way to use the default
values in postgres is to not supply those values (whether they be null
or otherwise) to the insert query - which sort of makes sense.
The way to disregard a value on insert in Slick is to set the type of
the column to O.AutoInc but when using forceInsert, that is overriden!

There was however a legit reason for using the forceInsert - namely when
I want to insert uuid from the client while still also supporting the
default values in cases where I do not supply a uuid. That worked fine
with mysql since forceInsert:ing null values is equivalent to letting
the default values set by the databse be used. But not so for
Postgres...